implementado camada de apresentacao

+++ DeveloperStore/DeveloperStore.Application/Services/ProductService.cs
@@ -48,9 +48,17 @@ namespace DeveloperStore.Application.Services
             try
             {
                 var model = _repo.GetById(id);
-                result.Valid = _repo.Delete(id);
-                result.Message = "Excluido com Sucesso.";
-                result.Data = model;
+                if (model == null)
+                {
+                    result.Valid = false;
+                    result.Message = "Objeto não localizado.";
+                }
+                else
+                {
+                    result.Valid = _repo.Delete(id);
+                    result.Message = "Excluido com Sucesso.";
+                    result.Data = model;
+                }
             }
             catch (Exception ex)
             {
@@ -68,9 +76,17 @@ namespace DeveloperStore.Application.Services
             try
             {
                 var model = _repo.GetById(id);
-                result.Valid = _repo.Delete(id);
-                result.Message = "Excluido com Sucesso.";
-                result.Data = model;
+                if (model == null)
+                {
+                    result.Valid = false;
+                    result.Message = "Objeto não localizado.";
+                }
+                else
+                {
+                    result.Valid = _repo.Delete(id);
+                    result.Message = "Excluido com Sucesso.";
+                    result.Data = model;
+                }
             }
             catch (Exception ex)
             {
@@ -88,9 +104,17 @@ namespace DeveloperStore.Application.Services
             try
             {
                 var obj = _repo.GetById(id);
-                result.Valid = true;
-                result.Message = "Sucesso.";
-                result.Data = obj;
+                if (obj == null)
+                {
+                    result.Valid = false;
+                    result.Message = "Objeto não localizado.";
+                }
+                else
+                {
+                    result.Valid = true;
+                    result.Message = "Sucesso.";
+                    result.Data = obj;
+                }
             }
             catch (Exception ex)
             {
@@ -108,9 +132,17 @@ namespace DeveloperStore.Application.Services
             try
             {
                 var obj = _repo.GetById(id);
-                result.Valid = true;
-                result.Message = "Sucesso.";
-                result.Data = obj;
+                if (obj == null)
+                {
+                    result.Valid = false;
+                    result.Message = "Objeto não localizado.";
+                }
+                else
+                {
+                    result.Valid = true;
+                    result.Message = "Sucesso.";
+                    result.Data = obj;
+                }
             }
             catch (Exception ex)
             {
@@ -143,7 +175,7 @@ namespace DeveloperStore.Application.Services
         }
 
 
-        public IContractResult ListAll()
+        public async Task<IContractResult> ListAll()
         {
             var result = new ContractResult();
             try
